refactor(core): app_factory を薄い委譲に変更し UI/WEB 工場を分離 (Step 10)
